Add symbols from Learn Plover Lesson 7: Non-Letters
When contacting Zach about the discrepency between the Plover default
dictionary and his list of symbols in Learn Plover, he said that the list was intended to be merged into Plover. I have done the work. I tested these in my own dictionary and they all work as expected.

This is the list of symbols added:
	(Will refer to prefix&suffix symbols as "sticky".)
	- Sticky Equals
	- Sticky Caret: ^
	- Sticky <
	- Sticky >
	- Sticky |
	- Sticky _
	- Prefix (
	- Suffix )
	- Prefix [
	- Suffix ]
	- Prefix {
	- Suffix }
	- Sticky @
	- Sticky *
	- Sticky \
	- &
	- "Hashtag", prefix # (will type #word and #2, etc.)

This is the list of entries removed:
	- Remove the suffix "-star" to make room for "*". If you can think of a better resolution, please suggest it and I can re-submit the pull request.
	- Remove "paren" for "("
	- Remove "Hear about" for "#"